==> Code/01_Acquisition_and_Preprocessing(house_for_sale).ipynb <==
## Coded by 
- Frank Zhao
- Gustavo Ferreira
- Yan Li

## Data aquiring date: 10/16/2020

# Import modules

import json
from pprint import pprint
import glob
from datetime import datetime
import requests
import pandas as pd

# Self-define functions

## function for realtor api requests

def realtor_api_request(sort = "relevance", city = "Philadelphia", limit = "200", offset = "0", state_code = "PA"):
    url = "https://realtor.p.rapidapi.com/properties/v2/list-for-sale"

    querystring = {
        "sort": sort,
        "city": city,
        "limit": limit,
        "offset": offset,
        "state_code": state_code
    }

    headers = {
        'x-rapidapi-host': "realtor.p.rapidapi.com",
        'x-rapidapi-key': "e2aef74c96msh10b708b155c5b67p1e6d19jsn4cb3a156a37d"
        }

    response = requests.request("GET", url, headers=headers, params=querystring)
    
    return response.json()

## convert response to pandas df

def process_response(response_json):
    '''
    This function is to convert each request result to a dataframe.
    
    1. create an empty list
    2. loop for each response and get details from key 'properties' 
    3. convert details to df
    4. append single df to list
    5. concat the list to one df
    
    '''
    
    # empty list
    convert_list=[]
#     if response_json['properties']: return pd.DataFrame([])
    # loop through each 'properties'
    for col in response_json['properties']:
        
        # convert details to dataframe
        single_df = pd.DataFrame.from_dict(col, orient='index').T
        
        # append to list
        convert_list.append(single_df)
        
    # concat to a whole df, null for missing vals
    return pd.concat(convert_list, axis = 0, ignore_index=True, sort=False)

# Get data

## run for each offset
'''
There are 200 instances in each request. Restricted by Rapidapi
'''
def response_to_df(sort = "relevance", city = "Philadelphia", limit = "200", offset = "0", state_code = "PA"):
    property_for_sale = realtor_api_request(sort = sort, city = city, limit = limit, offset = offset, state_code = state_code)
    df_properties_for_sale_raw = process_response(response_json = property_for_sale)
    pd.set_option('display.max_columns', None)
    return df_properties_for_sale_raw


# Export response to csv file

## export each response to csv file
'''
Since each request could get 200 instances, after the each request, we increase the offset 200.
So on the second run, we can get data from 201 - 400, theoratically.
However, the system will keep updating the data so we may either get the data overlapped or miss some data.
We will deal with duplicated data later.

Currently, the loop will crash when the instances in the last run is less than 200. 
'''

start = datetime.now() ## to see how long will it finish the run
i=1
for offset in range(0, 9562, 200):
    response_to_df(offset = offset).to_csv(f'test{i}.csv',encoding='utf-8', index=False)
    if i % 10 == 0:
        print(f'The {i}th Time')
    i += 1    
end = datetime.now() -start

print(f'Total time is: {end}')

# Data preprocessing

## import all csv to one df

path = r'E:\Projects\Drexel\DSCI 591 Capstone I\Projects\csv'
all_files = glob.glob(path + "/*.csv")

## empty list
file_list = []

for filename in all_files:
    df = pd.read_csv(filename, index_col=None, header=0)
    file_list.append(df)
    
combined_df = pd.concat(file_list, axis = 0, ignore_index=True)

combined_df

There are 9762 instances we achieved. However, there should be some duplicated instances. 

## remove duplicated instances
raw_data = combined_df.drop_duplicates(subset=['property_id'])
raw_data

## export to csv file
raw_data.to_csv('raw_data.csv', encoding='utf-8', index=False)

And now, we have all 9562 instances.

## Remove unuseful cols
processed_data = raw_data.drop(columns=[
    'rdc_web_url',
    'office',
    'thumbnail',
    'list_tracking',
    'open_houses',
    'plan_id',
    'new_plan',
    'quick_to_sell_days',
    'virtual_tour',
    'client_display_flags',
    'lead_forms',
    'listing_id',
    'mls'
    
])
processed_data

## Reset index

processed_data = processed_data.reset_index(drop= True)
processed_data

## Export to csv file

processed_data.to_csv('processed_data.csv',encoding='utf-8', index=False)

# Extra more info within cols

## Check NaN in each col

processed_data.isna().sum()

processed_data['prop_sub_type'].unique()

import numpy as np

## extra dic type strings from col
import ast

def str_to_dic(string):
    if string == None: pass
    else:
        return ast.literal_eval(string)
    

def xtr_info_from_dict(col, level = 0, keys = []):
    if level == 0:
        for i in keys:

            if i not in col.keys(): return None
            else:
                col = col[i]
        return col        
    
    else:
        for i in keys:

            if i not in col[level-1].keys(): return None
            else:
                col = col[level-1][i]
        return col

l = []
for i in range(len(processed_data['agents'])):
    if str(processed_data['agents'][i]) == 'nan':
        l.append(None)
    else:
        l.append(xtr_info_from_dict(str_to_dic(processed_data['agents'][i]), level=0, keys = ['id']))

## function for extracting and converting info to df
def info_to_df(col, level, keys, df, col_name):
    l = []
    for i in range(len(col)):
        if str(col[i]) == 'nan':
            l.append(np.nan)
        else:
            l.append(xtr_info_from_dict(str_to_dic(col[i]), level = level, keys = keys))
    
    # convert list to df
    new_df = pd.DataFrame(l, columns=col_name)
    return new_df

## function: append col to df

def append_df(col, level, keys, col_name, df):
    
    new_df = info_to_df(col, level, keys, df, col_name)
    
    return df.join(new_df)

There are some cols have details such `address`, `branding`, `agents`, `lot_size`, and `mls`. We need to discuss what features should we extract from them.

append_df(col = processed_data['agents'], level = 0, keys = ['id'], df = processed_data, col_name = ['agent_id']).head()

str_to_dic(processed_data['branding'][0])['listing_office']['list_item']['name']

len(processed_data['branding'])

## copy df
update_df = processed_data.copy()

extract = {'address':['city', 'line', 'postal_code', 'state_code', 'county', 'lat', 'lon', 'neighborhood_name'],
           'building_size':['size']
          }

for k in extract:
    for info in extract[k]:
        update_df = append_df(col = update_df[k], level = 0, keys = [info], df = update_df, col_name = [info])

## rename building size
update_df = update_df.rename(columns={'size':'buiding_size(sqft)'})

## get lot size
update_df = append_df(col = update_df['lot_size'], level = 0, keys = ['size'], df = update_df, col_name = ['lot_size(sqft)'])

## get agent id
update_df = append_df(col = update_df['agents'], level = 1, keys = ['id'], df = update_df, col_name = ['agent_id'])

## get agent name
update_df = append_df(col = update_df['agents'], level = 1, keys = ['name'], df = update_df, col_name = ['agent_name'])

## get branding office name
brand_name = []
for i in range(len(processed_data['branding'])):
    if str(processed_data['branding'][i]) == 'nan':
        brand_name.append(np.nan)
    else:
        brand_name.append(str_to_dic(processed_data['branding'][i])['listing_office']['list_item']['name'])
    
# convert list to df
brand_name = pd.DataFrame(brand_name, columns=['brand_name'])

update_df = update_df.join(brand_name)

## drop original cols
update_df.drop(columns=['address', 'building_size', 'agents', 'branding', 'lot_size'], inplace=True)

update_df.head()

update_df.isna().sum()

## export to csv
# update_df.to_csv('update_df.csv',encoding='utf-8', index=False)

update_df.shape
