==> Code/02_get_json(house_details).ipynb <==
## Coded by Frank Zhao
## Credit to Yan Li who helped get extra accounts for API access and acquired half of the whole data

import json
from pprint import pprint
import glob
from datetime import datetime
import requests
import pandas as pd

# Get the property_id

for_sale = pd.read_csv('./data/update_df.csv')
property_id = for_sale['property_id'].to_list()

len(property_id)

property_id[2395]

# Access API

'''
change your api_key
'''
AFKii3 = "c6516b212fmshb0ee7c91af586c7p115a7fjsn8af2978b9daa" #
AFKii4 = "882d9628cemsh733a2b4e13ea535p154ee3jsn60934db327d4" #
AFKii5 = "7bdd355104msheb3429d4f417979p145955jsn505f99a2e648" #
AFKii6 = "5484c1f90dmshd017ab6421de589p192b54jsnabe973197d3e" #
AFKii7 = "57f30851b4mshb1217f9f536c85fp18190djsnc2812d306e58" #
AFKii8 = "a9a3fac5a2msh7e3ad153d7b79efp1ac821jsne3d56a3a32f0" #
AFKii9 = "a203ac31ecmsh3c8c6b1f21b728bp1c245ejsn189446d70911" #
AFKii10 = "0f015bb6f6mshe37d52d3cd8e390p15c0c2jsn86fc0fbd6cad" #
AFKii11 = "b9d04f682amsh3fc522afdadf518p166725jsn9201db1d0564" #
AFKii12 = "e25b8b24c3msh58ccfad6e572e3cp17703ajsn88048ae4e5dd" #
AFKii13 = "072c33bf1cmsh842625b11a889efp19968ajsn83d404cae462"

## functions

def realtor_api_request(property_id):
    url = "https://realtor.p.rapidapi.com/properties/v2/detail"

    querystring = {"property_id":property_id}

    headers = {
        'x-rapidapi-host': "realtor.p.rapidapi.com",
        'x-rapidapi-key': AFKii13
        }

    response = requests.request("GET", url, headers=headers, params=querystring)
    
    return response.json()

detail1 = realtor_api_request(property_id[3150])
detail1

%%time
'''
1. just change the range, for safety, run 450 each time, I have set the first run for you.
2. change your directory folder, it would be great not change the file name
'''
for i in range(4500,4800): 
    proper_id = property_id[i]
    js = realtor_api_request(proper_id) ## get the json response
    
    ## save json file
    with open(f'.\data\house_hist_json\p4501_4800\house_{i+1}_{proper_id}.txt', 'w') as outfile:
        json.dump(js, outfile)
    if (i+1) % 10 == 0:
        print(f'The {i+1}th Time')
