==> Code/03_json_to_csv(property_detail).ipynb <==
## Coded By Frank Zhao

import json
from pprint import pprint
import glob
from datetime import datetime
import requests
import pandas as pd
import re

# load json file


## path
json_path =  r'.\data\house_hist_json'
folder = ["\p1_450", "\p451_900", "\p901_1350", "\p1351_1800", "\p1801_2250", "\p2251_2700"]
csv_path = r".\data\hist_price_csv"

all_files = glob.glob(path + folder[0] +"/*.txt")

# extract info


def process_response(response_json):
    '''
    This function is to convert each request result to a dataframe.
    
    1. create an empty list
    2. loop for each response and get details from key 'properties' 
    3. convert details to df
    4. append single df to list
    5. concat the list to one df
    
    '''
    
    # empty list
    convert_list=[]
#     if response_json['properties']: return pd.DataFrame([])
    # loop through each 'properties'
    
    if 'properties' not in response_json.keys():
        return pd.DataFrame(columns = ['event_name',
                                       'date',
                                       'price',
                                       'price_range_min',
                                       'price_range_max',
                                       'price_changed',
                                       'sqft',
                                       'datasource_name',
                                       'source'
                                      ])
    
    if 'property_history' not in response_json['properties'][0].keys():
        return pd.DataFrame(columns = ['event_name',
                                       'date',
                                       'price',
                                       'price_range_min',
                                       'price_range_max',
                                       'price_changed',
                                       'sqft',
                                       'datasource_name',
                                       'source'
                                      ])
        
    
    for col in response_json['properties'][0]['property_history']:
        
        # convert details to dataframe
        single_df = pd.DataFrame.from_dict(col, orient='index').T
        
        # append to list
        convert_list.append(single_df)
        
    # concat to a whole df, null for missing vals
    return pd.concat(convert_list, axis = 0, ignore_index=True, sort=False)

# Combine all together

## path
json_path =  r'.\data\house_hist_json'
folder = ["\p1_450", "\p451_900", "\p901_1350", "\p1351_1800",
          "\p1801_2250", "\p2251_2700", "\p2701_3150", '\p3151_3600',
          '\p3601_4050', '\p4051_4500', '\p4501_4800', '\p4801_9562'
         ]
csv_path = r".\data\hist_price_csv"

%%time
## 1. loop into folders
count = 1

for f in folder:
    print(f"Now in folder: {f}.")
    all_files = glob.glob(json_path + f +"/*.txt") 
    
    # 2. loop in each file
    for file in all_files: # string of directory
        
        # 3. get property id
        p_id = re.findall('(?<=_)[A-Z]\d+(?=.)', file)[0]
        
        # 4.load json file
        with open(file) as ff:
            js = json.load(ff)
            
        # 5. extract historical price
        hist_df = process_response(js)
        
        # 6. insert property id col to df
        hist_df.insert(loc=0, column='property_id', value=p_id)
        
        # 7. export to csv
        hist_df.to_csv(f'{csv_path}{f}\\hist_price_{count}_{p_id}.csv',encoding='utf-8', index=False)
        
        if count % 100 == 0:
            print(f"The {count}th run")
            
        count += 1
        
    

# combine all csv to 1 df

%%time
## 1. loop into folders
count = 1

file_list = []
for f in folder:
    print(f"Now in folder: {f}.")
    price_df = glob.glob(csv_path + f +"/*.csv") 
    
    
    # 2. loop in each file
    for filename in price_df: # string of directory
        
        df = pd.read_csv(filename, index_col=None, header=0)
        file_list.append(df)
 
        if count % 500 == 0:
            print(f"The {count}th run")
            
        count += 1

combined_df = pd.concat(file_list, axis = 0, ignore_index=True)

# combined_df.to_csv(f'.\data\hist_price_total\hist_price_all.csv',encoding='utf-8', index=False)

## part 2

%%time
## import all csv to one df
count = 1

path = r'.\data\hist_price_csv\p2'
all_files = glob.glob(path + "/*.csv")

## empty list
file_list = []

for filename in all_files:
    df = pd.read_csv(filename, index_col=None, header=0)
    file_list.append(df)
    
    if count % 200 ==0:
        print(f"The {count}th run")
    count += 1
    
hist_p2 = pd.concat(file_list, axis = 0, ignore_index=True)

hist_p2

# hist_p2.to_csv(f'.\data\hist_price_total\hist_price_part2.csv',encoding='utf-8', index=False)
