==> Code/04_Match_zip_from_detail.ipynb <==
## Coded by Frank Zhao

import json
from pprint import pprint
import glob
from datetime import datetime
import requests
import pandas as pd
import numpy as np

for_sale = pd.read_csv('./data/update_df.csv')
hist_p2 = pd.read_csv("E:\Projects\Drexel\DSCI 591 Capstone I\Projects\data\hist_price_total\hist_price_part2.csv")

for_sale.head()

## Part II

hist_p2.head()

# create empty df
hist_p2['postal_code'] = np.NaN

## assign values to hist_p2

for i in range(len(hist_p2['property_id'])):
    hist_p2.at[i, 'postal_code'] = zip_dict[hist_p2.iloc[i]['property_id']]


hist_p2.head()

hist_p2['postal_code'] = hist_p2['postal_code'].astype(int)

hist_p2['postal_code'].unique()

len(hist_p2['postal_code'])

# All files together

hist_all = pd.read_csv("E:\Projects\Drexel\DSCI 591 Capstone I\Projects\data\hist_price_total\hist_price_all.csv")

# create empty df
hist_all['postal_code'] = np.NaN

%%time
## assign values to hist_p2
for i in range(len(hist_all['property_id'])):
    hist_all.at[i, 'postal_code'] = zip_dict[hist_all.iloc[i]['property_id']]


hist_all['postal_code'] = hist_all['postal_code'].astype(int)

# hist_all.to_csv(f'.\data\hist_price_total\hist_price_with_zip.csv',encoding='utf-8', index=False)
